pt/code-reasoner-app/backend/requirements.txt
fastapi
uvicorn
pydantic
requests
google-generativeai
python-dotenv
